Rename shared 'games' directory to 'templates'
We are moving away from the ambiguous term "games" (since some stages
*are* games) and instead calling these templates.
