+++ frontend/src/shared/templates/chip_negotiation.ts
