+++ frontend/src/shared/templates/flipcard.ts
