+++ frontend/src/shared/templates/fruit_test.ts
