+++ frontend/src/shared/templates/lost_at_sea.ts
