+++ frontend/src/shared/templates/reality_tv_chat.ts
